refactor and pray
