+++ src/Components/Plots/Timeline/timelineConstants.ts
@@ -1,0 +1,32 @@
+//nodes
+export const NODE_RADIUS = 8;
+// edges
+export const EDGE_STROKE_COLOR = "#999";
+export const CURVE_SIZE = 15;
+export const EDGE_WIDTH = 2;
+// tooltip
+export const TOOLTIP_BG_COLOR = "#f4f4f4";
+export const TOOLTIP_BORDER = "1px solid #ccc";
+export const TOOLTIP_PADDING = "8px";
+export const TOOLTIP_BORDER_RADIUS = "4px";
+export const TOOLTIP_FONT = "var(--text-body-shorthand-medium)";
+export const TOOLTIP_MOUSEOVER_DUR = 200;
+export const TOOLTIP_MOUSEOUT_DUR = 500;
+// legends
+export const LEGEND_SIZE = 12;
+export const LEGEND_SYMBOL_SIZE = 60;
+export const LEGENDS_SPACING = "100px";
+export const LEGEND_TEXT_MARGIN = "10px";
+// misc
+export const DATE_LABEL_HEIGHT = 21;
+export const MARGIN = { top: 10, right: 0, bottom: 10, left: 150 };
+export const MIN_LABEL_SPACING = 40; 
+
+// buttons
+export const BUTTON_STYLE: React.CSSProperties = {
+    width: "120px", height: "40px", padding: "5px 10px",
+    cursor: "pointer", backgroundColor: "transparent",
+    border: "2px solid #eef", display: "flex",
+    alignItems: "center", justifyContent: "center",
+    transition: "background-color 0.3s",
+};
